requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.119.0
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.3.0
pillow_heif==1.1.1
pydantic==2.12.3
pydantic_core==2.41.4
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
torch==2.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
fastapi
uvicorn
torch
pillow
numpy
bcrypt
python-multipart
pillow-heif
requests